[INCOMPATIBLE] In CSS, ensure note titles are horizontally centered even if they span several lines.  It won't work in IE < 8 due to their limitations, so everything remains the same there.  In other browsers, this commit makes date precede title in note headers.  To correct that, all users should update their notes to place <div class="title"> before <div class="date">, or switch to table markup.  The alternatives are demostrated in notes.examples.html file.
Here is a sample script in Python to place "date" after "title":

    #!/usr/bin/env python3

    import sys
    import re

    input = sys.stdin.read()
    R = re.compile(r"""(<div class="date">.*?</div>)(\s*)(<div class="title">.*?</div>)""", re.DOTALL)
    output = R.sub(r"\3\2\1", input)
    print(output, end='')

NOTE: Do not use the script if you don't understand what it does!

NOTE: This is a sample script.  Before running it and replacing the
notes with its output, backup the notes!  Moreover, it's recommended to
check what the script have done, using a difference viewer, e.g.,
console tool called `diff` or a desktop program like Meld [1]

[1]: http://en.wikipedia.org/wiki/Meld_(software)